Add EMLEPC class methods as static methods of EMLEBase module.
