+++ emle/models/__init__.py
@@ -20,17 +20,11 @@
 # along with EMLE-Engine If not, see <http://www.gnu.org/licenses/>.
 ######################################################################
 
-# Note that this file is empty since EMLECalculator and Socket should
-# be directly imported from their respective sub-modules. This is to
-# avoid severe module import overheads when running the client code,
-# which requires no EMLE functionality.
-
 """
 Torch modules for EMLE calculations.
 """
 
 from ._emle_base import EMLEBase
-from ._emle_pc import EMLEPC
 from ._emle import EMLE
 from ._ani import ANI2xEMLE
 from ._mace import MACEEMLE
